Ok, now using memory ! Should be much more efficient (reading the entire file, putting it in RAM and progressing through it without making memory dubbles!! 64bits file too...ouch, hardcore

+++ arbre.h
@@ -38,7 +38,7 @@ extern "C" {
 	long long unsigned elt;
 } t_element;*/
 
-typedef int t_element;
+typedef long long unsigned t_element;
 
 typedef struct t_noeud {
 	//t_element valeur;
